Refactored routes, requests now dont need to pass userId.

+++ src/main/kotlin/com/nirmal/routes/CommentRoutes.kt
@@ -17,7 +17,6 @@ import io.ktor.server.routing.*
 
 fun Route.createComment(
     commentService: CommentService,
-    userService: UserService,
 ) {
     authenticate {
         post("/api/comment/create") {
@@ -25,37 +24,32 @@ fun Route.createComment(
                 call.respond(HttpStatusCode.BadRequest)
                 return@post
             }
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ) {
-                when(commentService.createComment(request)) {
-                    is CommentService.ValidationEvent.ErrorEmptyField-> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = false,
-                                message = ApiResponseMessages.FIELDS_BLANK
-                            )
+            when(commentService.createComment(request, call.userId)) {
+                is CommentService.ValidationEvent.ErrorEmptyField-> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = ApiResponseMessages.FIELDS_BLANK
                         )
-                    }
-                    is CommentService.ValidationEvent.ErrorCommentTooLong -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = false,
-                                message = ApiResponseMessages.COMMENT_TOO_LONG
-                            )
+                    )
+                }
+                is CommentService.ValidationEvent.ErrorCommentTooLong -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = ApiResponseMessages.COMMENT_TOO_LONG
                         )
-                    }
-                    is CommentService.ValidationEvent.Success -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = true
-                            )
+                    )
+                }
+                is CommentService.ValidationEvent.Success -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = true
                         )
-                    }
+                    )
                 }
             }
         }
@@ -84,7 +78,6 @@ fun Route.getCommentForPost(
 fun Route.deleteComment(
     commentService: CommentService,
     likeService: LikeService,
-    userService: UserService
 ) {
     authenticate {
         delete("/api/comment/delete") {
@@ -92,24 +85,34 @@ fun Route.deleteComment(
                 call.respond(HttpStatusCode.BadRequest)
                 return@delete
             }
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ) {
-                val deleted = commentService.deleteComment(request.commentId)
-                if (deleted) {
-                    likeService.deleteLikesForParent(request.commentId)
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(successful = true)
-                    )
-                } else {
-                    call.respond(
-                        HttpStatusCode.NotFound,
-                        BasicApiResponse(successful = false)
-                    )
-                }
+            val comment = commentService.getCommentById(request.commentId)
+            if (comment?.userId != call.userId) {
+//                println(comment?.userId)
+//                println(call.userId)
+//                call.respond(
+//                    HttpStatusCode.OK,
+//                    BasicApiResponse(
+//                        false,
+//                        call.userId + " " + comment?.comment
+//                    )
+//                )
+                call.respond(HttpStatusCode.Unauthorized)
+                return@delete
             }
+            val deleted = commentService.deleteComment(request.commentId)
+            if (deleted) {
+                likeService.deleteLikesForParent(request.commentId)
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(successful = true)
+                )
+            } else {
+                call.respond(
+                    HttpStatusCode.NotFound,
+                    BasicApiResponse(successful = false)
+                )
+            }
+
         }
     }
 }
