+++ src/main/kotlin/com/nirmal/routes/PostRoutes.kt
@@ -19,7 +19,6 @@ import io.ktor.server.routing.*
 
 fun Route.createPost(
     postService: PostService,
-    userService: UserService
 ) {
     authenticate {
         post("/api/post/create") {
@@ -27,27 +26,25 @@ fun Route.createPost(
                 call.respond(HttpStatusCode.BadRequest)
                 return@post
             }
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongToUserId,
-            ) {
-                val didUserExist = postService.createPostIfUserExists(request)
-                if (!didUserExist) {
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(
-                            successful = false,
-                            message = ApiResponseMessages.USER_NOT_FOUND
-                        )
+
+            val userId = call.userId
+
+            val didUserExist = postService.createPostIfUserExists(request, userId)
+            if (!didUserExist) {
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(
+                        successful = false,
+                        message = ApiResponseMessages.USER_NOT_FOUND
                     )
-                } else {
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(
-                            successful = true
-                        )
+                )
+            } else {
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(
+                        successful = true
                     )
-                }
+                )
             }
         }
     }
@@ -55,33 +52,22 @@ fun Route.createPost(
 
 fun Route.getPostByFollows(
     postService: PostService,
-    userService: UserService
 ) {
     authenticate {
         get("/api/post/get") {
-            val userId = call.parameters[QueryParams.PARAM_USER_ID] ?: kotlin.run {
-                call.respond(
-                    HttpStatusCode.BadRequest
-                )
-                return@get
-            }
+
             val page = call.parameters[QueryParams.PARAM_PAGE]?.toIntOrNull() ?: 0
             val pageSize = call.parameters[QueryParams.PARAM_PAGE_SIZE]?.toIntOrNull() ?: Constants.DEFAULT_POST_PAGE_SIZE
 
-            ifEmailBelongsToUser(
-                userId = userId,
-                userService::doesEmailBelongToUserId
-            ) {
-                val posts = postService.getPostByFollows(
-                    userId = userId,
-                    page = page,
-                    pageSize = pageSize
-                )
-                call.respond(
-                    HttpStatusCode.OK,
-                    posts
-                )
-            }
+            val posts = postService.getPostByFollows(
+                userId = call.userId,
+                page = page,
+                pageSize = pageSize
+            )
+            call.respond(
+                HttpStatusCode.OK,
+                posts
+            )
         }
     }
 }
@@ -89,30 +75,32 @@ fun Route.getPostByFollows(
 fun Route.deletePost(
     postService: PostService,
     likeService: LikeService,
-    userService: UserService
 ) {
-    delete("/api/post/delete") {
-        val request = call.receiveOrNull<DeletePostRequest>() ?: kotlin.run {
-            call.respond(
-                HttpStatusCode.BadRequest
-            )
-            return@delete
-        }
-        val post = postService.getPost(request.postId)
-        if (post == null) {
-            call.respond(HttpStatusCode.NotFound)
-            return@delete
-        }
-        ifEmailBelongsToUser(
-            userId = post.userId,
-            validateEmail = userService::doesEmailBelongToUserId
-        ) {
-            postService.deletePost(postId = request.postId)
-            likeService.deleteLikesForParent(request.postId)
+    authenticate {
+        delete("/api/post/delete") {
+            val request = call.receiveOrNull<DeletePostRequest>() ?: kotlin.run {
+                call.respond(
+                    HttpStatusCode.BadRequest
+                )
+                return@delete
+            }
+            val post = postService.getPost(request.postId)
+            if (post == null) {
+                call.respond(HttpStatusCode.NotFound)
+                return@delete
+            }
+            if (post.userId == call.userId) {
+                postService.deletePost(postId = request.postId)
+                likeService.deleteLikesForParent(request.postId)
                 // TODO: DELETE COMMENTS FROM POST
-            call.respond(HttpStatusCode.OK)
+                call.respond(HttpStatusCode.OK)
+            } else {
+                call.respond(HttpStatusCode.Unauthorized)
+            }
+
         }
     }
+
 }
